+++ Templates/Ie.php
@@ -1,0 +1,90 @@
+<?php
+/**
+ * Novutec Domain Tools
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @category   Novutec
+ * @package    DomainParser
+ * @copyright  Copyright (c) 2007 - 2013 Novutec Inc. (http://www.novutec.com)
+ * @license    http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+/**
+ * @namespace Novutec\WhoisParser
+ */
+namespace Novutec\WhoisParser;
+
+/**
+ * Template for .IE
+ *
+ * @category   Novutec
+ * @package    WhoisParser
+ * @copyright  Copyright (c) 2007 - 2013 Novutec Inc. (http://www.novutec.com)
+ * @license    http://www.apache.org/licenses/LICENSE-2.0
+ */
+class Template_Ie extends AbstractTemplate
+{
+
+    /**
+	 * Blocks within the raw output of the whois
+	 * 
+	 * @var array
+	 * @access protected
+	 */
+    protected $blocks = array(1 => '/domain:(?>[\x20\t]*)(.*?)(?=person)/is', 
+            2 => '/person:(?>[\x20\t]*).*?[\n]{2}/is');
+
+    /**
+	 * Items for each block
+	 * 
+	 * @var array
+	 * @access protected
+	 */
+    protected $blockItems = array(
+            1 => array('/^descr:(?>[\x20\t]*)(.+)$/im' => 'contacts:owner:address', 
+                    '/^admin-c:(?>[\x20\t]*)(.+)$/im' => 'network:contacts:admin', 
+                    '/^tech-c:(?>[\x20\t]*)(.+)$/im' => 'network:contacts:tech', 
+                    '/^status:(?>[\x20\t]*)(.+)$/im' => 'status', 
+                    '/^nserver:(?>[\x20\t]*)(.+)$/im' => 'nameserver', 
+                    '/^registration:(?>[\x20\t]*)(.+)$/im' => 'created', 
+                    '/^renewal:(?>[\x20\t]*)(.+)$/im' => 'expires'), 
+            
+            2 => array('/^nic-hdl:(?>[\x20\t]*)(.+)$/im' => 'contacts:handle', 
+                    '/^person:(?>[\x20\t]*)(.+)$/im' => 'contacts:name'));
+
+    /**
+     * RegEx to check availability of the domain name
+     *
+     * @var string
+     * @access protected
+     */
+    protected $available = '/Not Registered/i';
+
+    /**
+     * After parsing do something
+     *
+     * Fix owner name
+     *
+     * @param  object &$WhoisParser
+     * @return void
+     */
+    public function postProcess(&$WhoisParser)
+    {
+        $ResultSet = $WhoisParser->getResult();
+        
+        if (isset($ResultSet->contacts->owner[0]->address[0])) {
+            $ResultSet->contacts->owner[0]->name = $ResultSet->contacts->owner[0]->address[0];
+            $ResultSet->contacts->owner[0]->address = null;
+        }
+    }
+}